Revert "Change SelectionPercentage from double to int"
This reverts commit 1fa94ab94a29211062dd3f5a7ec6e0c2ea702ad7.

+++ DistLearning.Service/DTO/Responses/AnswerResponseWithStatistics.cs
@@ -3,5 +3,5 @@ namespace DistLearning.Service.DTO.Responses;
 public class AnswerResponseWithStatistics : BaseAnswerOptionResponse
 {
     public int SelectionCount { get; set; }
-    public int SelectionPercentage { get; set; }
+    public double SelectionPercentage { get; set; }
 }
